+++ test/sampledata.py
@@ -354,6 +354,7 @@ def create_cci_lccs_class_var(flag_values_as_list=False):
 
 
 class SourceDatasetMixin:
+
     @classmethod
     def new_2x2_dataset_with_irregular_coords(cls):
         lon = np.array([[1.0, 6.0], [0.0, 2.0]])
@@ -372,6 +373,40 @@ class SourceDatasetMixin:
         lon = np.array([[+179.0, -176.0], [+178.0, +180.0]])
         lat = np.array([[56.0, 53.0], [52.0, 50.0]])
         rad = np.array([[1.0, 2.0], [3.0, 4.0]])
+        return xr.Dataset(
+            dict(
+                lon=xr.DataArray(lon, dims=("y", "x")),
+                lat=xr.DataArray(lat, dims=("y", "x")),
+                rad=xr.DataArray(rad, dims=("y", "x")),
+            )
+        )
+
+    @classmethod
+    def new_4x4_dataset_with_irregular_coords(cls):
+        lon = np.array(
+            [
+                [1.0, 2.0, 3.0, 4.0],
+                [0.0, 1.0, 2.0, 3.0],
+                [-1.0, 0.0, 1.0, 2.0],
+                [-2.0, -1.0, 0.0, 1.0],
+            ]
+        )
+        lat = np.array(
+            [
+                [56.0, 55.0, 54.0, 53.0],
+                [55.0, 54.0, 53.0, 52.0],
+                [54.0, 53.0, 52.0, 51.0],
+                [53.0, 52.0, 51.0, 50.0],
+            ]
+        )
+        rad = np.array(
+            [
+                [1.0, 2.0, 3.0, 4.0],
+                [5.0, 6.0, 7.0, 8.0],
+                [9.0, 10.0, 11.0, 12.0],
+                [13.0, 14.0, 15.0, 16.0],
+            ]
+        )
         return xr.Dataset(
             dict(
                 lon=xr.DataArray(lon, dims=("y", "x")),