+++ xcube/core/gridmapping/coords.py
@@ -19,6 +19,8 @@ from .helpers import _assert_valid_xy_names
 from .helpers import _default_xy_var_names
 from .helpers import _normalize_crs
 from .helpers import _normalize_int_pair
+from .helpers import _normalize_number_pair
+from .helpers import Number
 from .helpers import _to_int_or_float
 from .helpers import from_lon_360
 from .helpers import round_to_fraction
@@ -61,10 +63,7 @@ class Coords1DGridMapping(CoordsGridMapping):
         y, x = xr.broadcast(self._y_coords, self._x_coords)
         tmp = xr.concat([x, y], dim="coord")
         return tmp.chunk(
-            {
-                dim: size for (dim, size) in
-                zip(tmp.dims, self.xy_coords_chunks)
-            }
+            {dim: size for (dim, size) in zip(tmp.dims, self.xy_coords_chunks)}
         )
 
 
@@ -76,10 +75,7 @@ class Coords2DGridMapping(CoordsGridMapping):
     def _new_xy_coords(self) -> xr.DataArray:
         tmp = xr.concat([self._x_coords, self._y_coords], dim="coord")
         return tmp.chunk(
-            {
-                dim: size for (dim, size) in
-                zip(tmp.dims, self.xy_coords_chunks)
-            }
+            {dim: size for (dim, size) in zip(tmp.dims, self.xy_coords_chunks)}
         )
 
 
@@ -88,6 +84,7 @@ def new_grid_mapping_from_coords(
     y_coords: xr.DataArray,
     crs: Union[str, pyproj.crs.CRS],
     *,
+    xy_res: Union[Number, tuple[Number, Number]] = None,
     tile_size: Union[int, tuple[int, int]] = None,
     tolerance: float = DEFAULT_TOLERANCE,
 ) -> GridMapping:
@@ -108,10 +105,7 @@ def new_grid_mapping_from_coords(
     tile_size = _normalize_int_pair(tile_size, default=None)
     is_lon_360 = None  # None means "not yet known"
     if crs.is_geographic:
-        is_lon_360 = bool(np.any(x_coords > 180))
-
-    x_res = 0
-    y_res = 0
+        is_lon_360 = bool(da.any(x_coords > 180))
 
     if x_coords.ndim == 1:
         # We have 1D x,y coordinates
@@ -136,16 +130,20 @@ def new_grid_mapping_from_coords(
                 x_diff = _abs_no_zero(x_coords.diff(dim=x_dim))
                 is_lon_360 = True
 
-        x_res, y_res = x_diff[0], y_diff[0]
-        x_diff_equal = np.allclose(x_diff, x_res, atol=tolerance)
-        y_diff_equal = np.allclose(y_diff, y_res, atol=tolerance)
-        is_regular = x_diff_equal and y_diff_equal
-        if is_regular:
-            x_res = round_to_fraction(x_res, 5, 0.25)
-            y_res = round_to_fraction(y_res, 5, 0.25)
+        if xy_res is not None:
+            x_res, y_res = _normalize_number_pair(xy_res)
         else:
-            x_res = round_to_fraction(float(np.nanmedian(x_diff)), 2, 0.5)
-            y_res = round_to_fraction(float(np.nanmedian(y_diff)), 2, 0.5)
+            x_res = x_diff[0]
+            y_res = y_diff[0]
+            is_regular = da.allclose(x_diff, x_res, atol=tolerance) and da.allclose(
+                x_diff, y_res, atol=tolerance
+            )
+            if is_regular:
+                x_res = round_to_fraction(x_res, 5, 0.25)
+                y_res = round_to_fraction(y_res, 5, 0.25)
+            else:
+                x_res = round_to_fraction(float(np.nanmedian(x_diff)), 2, 0.5)
+                y_res = round_to_fraction(float(np.nanmedian(y_diff)), 2, 0.5)
 
         if (
             tile_size is None
@@ -178,10 +176,10 @@ def new_grid_mapping_from_coords(
         x = da.asarray(x_coords)
         y = da.asarray(y_coords)
 
-        x_x_diff = _abs_no_nan(da.diff(x, axis=1))
-        x_y_diff = _abs_no_nan(da.diff(x, axis=0))
-        y_x_diff = _abs_no_nan(da.diff(y, axis=1))
-        y_y_diff = _abs_no_nan(da.diff(y, axis=0))
+        x_x_diff = _abs_no_nan(da.diff(x[0, :]))
+        x_y_diff = _abs_no_nan(da.diff(x[:, 0]))
+        y_x_diff = _abs_no_nan(da.diff(y[0, :]))
+        y_y_diff = _abs_no_nan(da.diff(y[:, 0]))
 
         if not is_lon_360 and crs.is_geographic:
             is_anti_meridian_crossed = da.any(da.max(x_x_diff) > 180) or da.any(
@@ -190,23 +188,40 @@ def new_grid_mapping_from_coords(
             if is_anti_meridian_crossed:
                 x_coords = to_lon_360(x_coords)
                 x = da.asarray(x_coords)
-                x_x_diff = _abs_no_nan(da.diff(x, axis=1))
-                x_y_diff = _abs_no_nan(da.diff(x, axis=0))
+                x_x_diff = _abs_no_nan(da.diff(x[0, :]))
+                x_y_diff = _abs_no_nan(da.diff(x[:, 0]))
                 is_lon_360 = True
 
-        is_regular = False
+        if xy_res is not None:
+            x_res, y_res = _normalize_number_pair(xy_res)
+        else:
+            x_res = x_x_diff[0]
+            y_res = y_y_diff[0]
 
-        if da.all(x_y_diff == 0) and da.all(y_x_diff == 0):
-            x_res = x_x_diff[0, 0]
-            y_res = y_y_diff[0, 0]
-            is_regular = (
-                da.allclose(x_x_diff[0, :], x_res, atol=tolerance)
-                and da.allclose(x_x_diff[-1, :], x_res, atol=tolerance)
-                and da.allclose(y_y_diff[:, 0], y_res, atol=tolerance)
-                and da.allclose(y_y_diff[:, -1], y_res, atol=tolerance)
-            )
+        is_regular = (
+            da.allclose(x_x_diff, x_res, atol=tolerance)
+            and da.allclose(y_y_diff, y_res, atol=tolerance)
+            and da.all(x_y_diff == 0)
+            and da.all(y_x_diff == 0)
+        )
 
-        if not is_regular:
+        if not is_regular and xy_res is None:
+            x_x_diff = _abs_no_nan(da.diff(x, axis=1))
+            x_y_diff = _abs_no_nan(da.diff(x, axis=0))
+            y_x_diff = _abs_no_nan(da.diff(y, axis=1))
+            y_y_diff = _abs_no_nan(da.diff(y, axis=0))
+
+            if not is_lon_360 and crs.is_geographic:
+                is_anti_meridian_crossed = da.any(da.max(x_x_diff) > 180) or da.any(
+                    da.max(x_y_diff) > 180
+                )
+                if is_anti_meridian_crossed:
+                    x_coords = to_lon_360(x_coords)
+                    x = da.asarray(x_coords)
+                    y_x_diff = _abs_no_nan(da.diff(y, axis=1))
+                    y_y_diff = _abs_no_nan(da.diff(y, axis=0))
+                    is_lon_360 = True
+
             # Let diff arrays have same shape as original by
             # doubling last rows and columns.
             x_x_diff_c = da.concatenate([x_x_diff, x_x_diff[:, -1:]], axis=1)
@@ -247,17 +262,21 @@ def new_grid_mapping_from_coords(
 
         if tile_size is not None:
             tile_width, tile_height = tile_size
-            x_coords = x_coords.chunk({
+            x_coords = x_coords.chunk(
+                {
                     x_coords.dims[0]: tile_height,
-                    x_coords.dims[1]: tile_height,
-            })
-            y_coords = y_coords.chunk({
+                    x_coords.dims[1]: tile_width,
+                }
+            )
+            y_coords = y_coords.chunk(
+                {
                     y_coords.dims[0]: tile_height,
-                    y_coords.dims[1]: tile_height,
-            })
+                    y_coords.dims[1]: tile_width,
+                }
+            )
 
         # Guess j axis direction
-        is_j_axis_up = np.all(y_coords[0, :] < y_coords[-1, :]) or None
+        is_j_axis_up = da.all(y_coords[0, :] < y_coords[-1, :]) or None
 
     assert_true(
         x_res > 0 and y_res > 0,