+++ xcube/core/resampling/spatial.py
@@ -176,27 +176,15 @@ def resample_in_space(
                 gm_name=gm_name,
                 **(rectify_kwargs or {}),
             )
-
-        # Source has higher resolution than target.
-        # Downscale first, then rectify
-        if source_gm.is_regular:
-            # If source is regular
-            downscaled_gm = source_gm.scale((x_scale, y_scale))
-            downscaled_dataset = resample_dataset(
-                source_ds,
-                ((x_scale, 1, 0), (1, y_scale, 0)),
-                size=downscaled_gm.size,
-                tile_size=source_gm.tile_size,
-                xy_dim_names=source_gm.xy_dim_names,
-                var_configs=var_configs,
-            )
         else:
+            # Source has higher resolution than target.
+            # Downscale first, then rectify
             _, downscaled_size = scale_xy_res_and_size(
                 source_gm.xy_res, source_gm.size, (x_scale, y_scale)
             )
             downscaled_dataset = resample_dataset(
                 source_ds,
-                ((x_scale, 1, 0), (1, y_scale, 0)),
+                ((1 / x_scale, 0, 0), (0, 1 / y_scale, 0)),
                 size=downscaled_size,
                 tile_size=source_gm.tile_size,
                 xy_dim_names=source_gm.xy_dim_names,
@@ -205,21 +193,21 @@ def resample_in_space(
             downscaled_gm = GridMapping.from_dataset(
                 downscaled_dataset,
                 tile_size=source_gm.tile_size,
-                prefer_crs=source_gm.crs,
+                crs=source_gm.crs,
             )
-        return rectify_dataset(
-            downscaled_dataset,
-            source_gm=downscaled_gm,
-            ref_ds=ref_ds,
-            target_gm=target_gm,
-            encode_cf=encode_cf,
-            gm_name=gm_name,
-            **(rectify_kwargs or {}),
-        )
+            return rectify_dataset(
+                downscaled_dataset,
+                source_gm=downscaled_gm,
+                ref_ds=ref_ds,
+                target_gm=target_gm,
+                encode_cf=encode_cf,
+                gm_name=gm_name,
+                **(rectify_kwargs or {}),
+            )
 
     # If CRSes are not both geographic and their CRSes are different
     # transform the source_gm so its CRS matches the target CRS:
-    transformed_source_gm = source_gm.transform(target_gm.crs)
+    transformed_source_gm = source_gm.transform(target_gm.crs, xy_res=target_gm.xy_res)
     transformed_x, transformed_y = transformed_source_gm.xy_coords
     return resample_in_space(
         source_ds.assign(transformed_x=transformed_x, transformed_y=transformed_y),