+++ CHANGES.md
@@ -1,5 +1,20 @@
 ## Changes in 1.7.2 (in development)
 
+### Enhancements
+
+* A `xy_res` keyword argument was added to the `transform()` method of
+  `xcube.core.gridmapping.GridMapping`, enabling users to set the grid-mapping 
+  resolution directly, which speeds up the method by avoiding time-consuming 
+  spatial resolution estimation. (#1082)
+* The behaviour of the function `xcube.core.resample.resample_in_space()` has 
+  been changed if no `tile_size` is specified for the target grid mapping. It now 
+  defaults to the `tile_size` of the source grid mapping, improving the 
+  user-friendliness of resampling and reprojection.
+
+### Fixes
+
+* The function `xcube.core.resample.resample_in_space()` now always operates
+   lazily and therefore supports chunk-wise, parallel processing. (#1079)
 
 ## Changes in 1.7.1